chore: macro refactoring
