+++ crates/mqtt-macros/src/entity.rs
@@ -1,217 +1,122 @@
-mod entity_struct;
-mod input;
-mod invalidity;
-mod validate;
+use crate::{json_doc::DocumentStruct, util::Prepend};
+use darling::FromDeriveInput;
+use proc_macro2::TokenStream;
+use quote::{quote, ToTokens};
+use syn::{parse2, FieldsNamed};
 
-use convert_case::{Case, Casing};
-use darling::{error::Accumulator, usage::GenericsExt, Error, Result};
-use proc_macro2::{Span, TokenStream};
-use quote::{format_ident, ToTokens};
-use syn::parse2;
+fn common_fields() -> FieldsNamed {
+	let tokens = quote! {{
+		/// A list of MQTT topics subscribed to receive availability (online/offline) updates.
+		#[serde(borrow, default, skip_serializing_if = "<[crate::availability::Availability]>::is_empty")]
+		#[entity(validate)]
+		pub availability: ::std::borrow::Cow<'a, [crate::availability::Availability<'a>]>,
 
-use self::input::{AdditionalInvalidities, AdditionalProps};
+		/// When `availability` is configured, this controls the conditions needed
+		/// to set the entity to `available`.
+		#[serde(default, skip_serializing_if = "crate::availability::AvailabilityMode::is_default")]
+		pub availability_mode: crate::availability::AvailabilityMode,
 
-struct EntityStruct {
-	ident: syn::Ident,
-	invalidity_ident: syn::Ident,
-	vis: syn::Visibility,
-	generics: syn::Generics,
-	docs: Vec<syn::Attribute>,
-	attrs: Vec<syn::Attribute>,
-	fields: Vec<EntityField>,
-	additional_invalidities: Option<AdditionalInvalidities>,
-	additional_props: Option<AdditionalProps>,
+		/// Information about the device this entity is a part of to tie it into the device registry.
+		/// Only works through MQTT discovery and when `unique_id` is set.
+		/// At least one of identifiers or connections must be present to identify the device.
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		#[entity(validate)]
+		pub device: Option<crate::device::Device<'a>>,
+
+		/// Flag which defines if the entity should be enabled when first added.
+		/// Defaults to `true`.
+		#[serde(default, skip_serializing_if = "Option::is_none")]
+		pub enabled_by_default: Option<bool>,
+
+		/// The encoding of the payloads received and published messages. Set to "" to disable decoding of incoming payload.
+		/// Defaults to `"utf-8"`.
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		pub encoding: Option<::std::borrow::Cow<'a, str>>,
+
+		/// The [category] of the entity.
+		///
+		/// [category]: https://developers.home-assistant.io/docs/core/entity#generic-properties
+		#[serde(default, skip_serializing_if = "crate::entity_category::EntityCategory::is_none")]
+		pub entity_category: crate::entity_category::EntityCategory,
+
+		/// [Icon][icon] for the entity.
+		///
+		/// [icon]: https://www.home-assistant.io/docs/configuration/customizing-devices/#icon
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		#[entity(validate)]
+		pub icon: Option<crate::icon::Icon<'a>>,
+
+		/// Defines a [template][template] to extract the JSON dictionary from messages received
+		/// on the `json_attributes_topic`.
+		///
+		/// [template]: https://www.home-assistant.io/docs/configuration/templating/#processing-incoming-data
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		#[entity(validate)]
+		pub json_attributes_template: Option<crate::template::Template<'a>>,
+
+		/// The MQTT topic subscribed to receive a JSON dictionary payload and then set as entity
+		/// attributes.
+		///
+		/// Implies `force_update` of the current state when a message is received on this topic.
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		#[entity(validate)]
+		pub json_attributes_topic: Option<crate::topic::Topic<'a>>,
+
+		/// The name of the MQTT entity.
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		#[entity(validate)]
+		pub name: Option<crate::name::Name<'a>>,
+
+		/// Used instead of `name` for automatic generation of `entity_id`.
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		pub object_id: Option<::std::borrow::Cow<'a, str>>,
+
+		/// The maximum QoS level of the state topic.
+		#[serde(default, skip_serializing_if = "crate::qos::MqttQoS::is_default")]
+		pub qos: crate::qos::MqttQoS,
+
+		/// An ID that uniquely identifies this entity. If two entities have the same unique ID,
+		/// Home Assistant will raise an exception.
+		#[serde(borrow, default, skip_serializing_if = "Option::is_none")]
+		#[entity(validate)]
+		pub unique_id: Option<crate::unique_id::UniqueId<'a>>,
+	}};
+
+	let fields: FieldsNamed = syn::parse2(tokens).unwrap();
+	fields
 }
 
-impl EntityStruct {
-	fn entity_struct(&self) -> impl ToTokens + '_ {
-		entity_struct::entity_struct(self)
-	}
+struct EntityStruct(DocumentStruct);
 
-	fn invalidity_enum(&self) -> impl ToTokens + '_ {
-		invalidity::invalidity_enum(self)
-	}
+impl FromDeriveInput for EntityStruct {
+	fn from_derive_input(input: &syn::DeriveInput) -> darling::Result<Self> {
+		let mut input = input.clone();
 
-	fn validate(&self) -> impl ToTokens + '_ {
-		validate::validation(self)
-	}
-}
-
-impl TryFrom<input::EntityStructInput> for EntityStruct {
-	type Error = darling::Error;
-
-	fn try_from(value: input::EntityStructInput) -> Result<Self> {
-		let mut accumulator = Accumulator::default();
+		if let syn::Data::Struct(syn::DataStruct {
+			fields: syn::Fields::Named(f),
+			..
+		}) = &mut input.data
 		{
-			let lifetimes = value.generics.declared_lifetimes();
-			let mut has_a = false;
-			let mut first = true;
-			for lifetime in lifetimes {
-				if first {
-					first = false;
-				}
-
-				if lifetime.ident == "a" {
-					has_a = true;
-				} else {
-					accumulator.push(
-						Error::custom("Entities must only have a single lifetime named 'a")
-							.with_span(&lifetime.ident),
-					);
-				}
-			}
-
-			if !has_a {
-				let error = Error::custom("Entities must have a lifetime 'a");
-				let error = if first {
-					error.with_span(&value.ident)
-				} else {
-					error.with_span(&value.generics)
-				};
-				accumulator.push(error);
-			}
+			f.prepend(common_fields());
 		}
 
-		let fields = match value.data.take_struct() {
-			None => {
-				accumulator.push(Error::custom("Entities must be structs").with_span(&value.ident));
-				Vec::new()
-			}
-			Some(data) if !data.is_struct() => {
-				accumulator.push(Error::custom("Entities must be structs").with_span(&value.ident));
-				Vec::new()
-			}
-			Some(data) => {
-				let mut fields = Vec::with_capacity(data.len());
-				for field in data {
-					if let Some(field) = accumulator.handle(EntityField::try_from(field)) {
-						fields.push(field);
-					}
-				}
-
-				fields
-			}
-		};
-
-		let (docs, attrs): (Vec<_>, Vec<_>) = value
-			.attrs
-			.into_iter()
-			.partition(|attr| attr.path.is_ident("doc"));
-
-		let invalidity_ident = format_ident!("{}Invalidity", &value.ident, span = Span::call_site());
-
-		accumulator.finish_with(Self {
-			ident: value.ident,
-			invalidity_ident,
-			vis: value.vis,
-			generics: value.generics,
-			fields,
-			docs,
-			attrs,
-			additional_invalidities: value.validate,
-			additional_props: value.extend_json,
-		})
+		DocumentStruct::from_derive_input(&input).map(Self)
 	}
 }
 
 impl ToTokens for EntityStruct {
 	fn to_tokens(&self, tokens: &mut TokenStream) {
-		self.entity_struct().to_tokens(tokens);
-		self.invalidity_enum().to_tokens(tokens);
-		self.validate().to_tokens(tokens);
+		self.0.document_struct().to_tokens(tokens);
+		self.0.ctor().to_tokens(tokens);
+		self.0.builders().to_tokens(tokens);
+		self.0.invalidity_enum().to_tokens(tokens);
+		self.0.validate().to_tokens(tokens);
+		self.0.serde().to_tokens(tokens);
 	}
 }
 
-struct EntityField {
-	ident: syn::Ident,
-	variant_ident: syn::Ident,
-	ty: syn::Type,
-	docs: Vec<syn::Attribute>,
-	attrs: Vec<syn::Attribute>,
-	validate: FieldValidation,
-	required: bool,
-}
-
-enum FieldValidation {
-	None,
-	Default,
-	With(syn::Path),
-}
-impl FieldValidation {
-	fn then<R>(&self, f: impl FnOnce(Option<&syn::Path>) -> R) -> Option<R> {
-		match self {
-			Self::None => None,
-			Self::Default => Some(f(None)),
-			Self::With(path) => Some(f(Some(path))),
-		}
-	}
-}
-
-impl From<input::FieldValidation> for FieldValidation {
-	fn from(value: input::FieldValidation) -> Self {
-		match value {
-			input::FieldValidation::None => Self::None,
-			input::FieldValidation::Default(_) => Self::Default,
-			input::FieldValidation::With(_, path) => Self::With(path),
-		}
-	}
-}
-
-impl TryFrom<input::EntityFieldInput> for EntityField {
-	type Error = darling::Error;
-
-	fn try_from(value: input::EntityFieldInput) -> Result<Self> {
-		let mut accumulator = Accumulator::default();
-		let ident = match value.ident {
-			Some(ident) => ident,
-			None => {
-				accumulator.push(Error::custom("Entity fields must be named"));
-				syn::Ident::new("unknown", proc_macro2::Span::call_site())
-			}
-		};
-
-		if !matches!(value.vis, syn::Visibility::Public(_)) {
-			accumulator.push(Error::custom("Entity fields must be public").with_span(&ident));
-		}
-
-		let ty = value.ty;
-		let (docs, attrs): (Vec<_>, Vec<_>) = value
-			.attrs
-			.into_iter()
-			.partition(|attr| attr.path.is_ident("doc"));
-		let validate = value.validate.into();
-		let has_default = attrs
-			.iter()
-			.any(|attr| attr.path.is_ident("serde") && attr.tokens.to_string().contains("default"));
-		let required = !has_default;
-
-		let variant_ident = format_ident!(
-			"{}",
-			ident
-				.to_string()
-				.from_case(Case::Snake)
-				.to_case(Case::Pascal),
-			span = Span::call_site(),
-		);
-
-		accumulator.finish_with(Self {
-			ident,
-			variant_ident,
-			ty,
-			attrs,
-			docs,
-			validate,
-			required,
-		})
-	}
-}
-
-// pub fn derive(input: syn::DeriveInput) -> Result<proc_macro2::TokenStream, darling::Error> {
-//   todo!()
-// }
-
-pub fn create(input: TokenStream) -> Result<TokenStream> {
+pub fn create(input: TokenStream) -> darling::Result<TokenStream> {
 	let parsed: syn::DeriveInput = parse2(input)?;
-	let entity = input::from_derive_input(&parsed)?;
-	EntityStruct::try_from(entity).map(ToTokens::into_token_stream)
+	let entity = EntityStruct::from_derive_input(&parsed)?;
+	Ok(entity.into_token_stream())
 }
