+++ crates/mqtt-macros/src/lib.rs
@@ -1,4 +1,5 @@
 mod entity;
+mod json_doc;
 mod util;
 
 use proc_macro::TokenStream;
@@ -10,3 +11,11 @@ pub fn entity_document(_attr: TokenStream, item: TokenStream) -> TokenStream {
 		Err(err) => err.write_errors().into(),
 	}
 }
+
+// #[proc_macro_attribute]
+// pub fn json_document(_attr: TokenStream, item: TokenStream) -> TokenStream {
+// 	match json_doc::create(item.into()) {
+// 		Ok(tokens) => tokens.into(),
+// 		Err(err) => err.write_errors().into(),
+// 	}
+// }
